(refactor) Move themes to `Light` and `Dark` components
